requirements.txt
streamlit==1.32.2
tensorflow==2.15.0
pillow==10.1.0
numpy==1.26.3